+++ src/main/java/net/uptheinter/interceptify/annotations/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package contains the annotations that you should apply to
+ * your own classes and methods for runtime interception.
+ */
+package net.uptheinter.interceptify.annotations;
