+++ src/main/java/net/uptheinter/interceptify/interfaces/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This package contains interfaces you should implement
+ * and pass to
+ * {@link net.uptheinter.interceptify.EntryPoint#entryPoint(net.uptheinter.interceptify.interfaces.StartupConfig,
+ *                                                          java.lang.String[]) EntryPoint.entryPoint}
+ */
+package net.uptheinter.interceptify.interfaces;
