+++ src/main/java/net/uptheinter/interceptify/internal/RuntimeHook.java
@@ -1,6 +1,7 @@
 package net.uptheinter.interceptify.internal;
 
 import net.bytebuddy.ByteBuddy;
+import net.bytebuddy.ClassFileVersion;
 import net.uptheinter.interceptify.EntryPoint;
 import net.uptheinter.interceptify.interfaces.StartupConfig;
 import net.uptheinter.interceptify.util.JarFiles;
@@ -12,6 +13,10 @@ import java.net.URL;
  * This is the primary point at which intercepting logic kicks off.
  * The class doesn't do all that much now but exists for the purposes
  * of facilitating Java Agent functionality in the future.
+ *
+ * As should be obvious from the package name, this class is internal.
+ * Use EntryPoint for your needs. This class is not subject to semantic
+ * versioning rules as you shouldn't be using it.
  */
 public final class RuntimeHook {
     private static ClassInjector ci;
@@ -30,6 +35,7 @@ public final class RuntimeHook {
         ci
             .setClassPath(classpaths)
             .defineMakePublicList(cfg.makePublic())
+            .defineMakePublicPredicate(cfg::shouldMakePublic)
             .collectMetadataFrom(jarFiles)
             .applyAnnotationsAndIntercept();
     }
@@ -43,6 +49,9 @@ public final class RuntimeHook {
      */
     @SuppressWarnings("unused")
     public static void premain(String args, Instrumentation instr) {
-        ci = new ClassInjector(new ByteBuddy(), instr);
+        var bb = new ByteBuddy();
+        ci = new ClassInjector(args == null || args.isEmpty() ?
+                bb :
+                bb.with(ClassFileVersion.ofJavaVersionString(args)), instr);
     }
 }
