+++ src/main/java/net/uptheinter/interceptify/internal/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Everything under this package is considered internal to
+ * Interceptify. No semantic versioning rules apply to anything
+ * under here and may be changed arbitrarily without notice since
+ * you shouldn't be using anything in here directly.
+ */
+package net.uptheinter.interceptify.internal;
