+++ src/main/java/net/uptheinter/interceptify/util/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This package contains an assortment of convenience
+ * functions which are primarily intended for the benefit
+ * of Interceptify - however, they are considered public
+ * and consequently <b>will</b> honour semantic versioning.
+ */
+package net.uptheinter.interceptify.util;
